requirements.txt
accelerate==0.23.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
async-timeout==5.0.1
attrs==24.3.0
av==14.0.1
bitsandbytes==0.41.0
blessed==1.20.0
cachetools==5.5.0
click==8.1.8
cmake==3.31.2
contourpy==1.3.1
cycler==0.12.1
datasets==3.2.0
decord==0.6.0
deepspeed==0.12.2
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.6.1
einops-exts==0.0.4
exceptiongroup==1.2.2
fire==0.7.0
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.9.0
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.43
gpustat==1.1.1
h11==0.14.0
hjson==3.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.0
idna==3.10
iopath==0.1.10
joblib==1.4.2
kiwisolver==1.4.8
lit==18.1.8
logzero==1.7.0
markdown-it-py==3.0.0
matplotlib==3.10.0
mdurl==0.1.2
mkl-service==2.4.0
multidict==6.1.0
multiprocess==0.70.16
ninja==1.11.1.3
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==12.560.30
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
nvitop==1.4.0
openai==1.4.0
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
parameterized==0.9.0
peft==0.5.0
platformdirs==4.3.6
portalocker==3.0.0
propcache==0.2.1
protobuf==5.29.2
psutil==6.1.1
py-cpuinfo==9.0.0
pyarrow==18.1.0
pybind11==2.13.6
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
pynvml==12.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytorchvideo==0.1.5
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
rich==13.9.4
safetensors==0.4.5
scikit-learn==1.2.2
scipy==1.14.1
sentencepiece==0.1.99
sentry-sdk==2.19.2
setproctitle==1.3.4
shortuuid==1.0.13
shtab==1.7.1
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
tabulate==0.9.0
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
timm==0.6.13
tokenizers==0.15.0
tqdm==4.67.1
transformers==4.36.2
transformers-stream-generator==0.0.5
triton==2.0.0
tyro==0.7.2
tzdata==2024.2
wandb==0.19.1
wcwidth==0.2.13
xxhash==3.5.0
yacs==0.1.8
yarl==1.18.3
